LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting-api/reporting-api-sends-reports-on-violation.https.sub-expected.txt



PASS Test that image does not load
PASS Event is fired
FAIL Report is observable to ReportingObserver Can't find variable: ReportingObserver
FAIL Violation report status OK. undefined is not an object (evaluating 'data[0]["body"]')
